Add `PredictionService` to replace `PredictionRunner`
This adds a simpler `PredictionService` to replace the existing
`PredictionRunner`.

Unlike `PredictionRunner`, the new service doesn't need any auxiliary
threads to keep track of the state of the running setup or prediction
jobs. Instead, it manages the lifecycle of `PredictionEventHandler` and
(new) `SetupEventHandler` objects which aggregate the individual events
emitted by `Worker` into a result object for setups and predictions
respectively.

We expect that this code will change significantly in the near future,
as it will need to keep track of multiple predictions at once.

Co-Authored-By: Nick Stenning <nick@whiteink.com>