Replace `PredictionRunner` with `PredictionService`
The fun bit! This rips out the existing `PredictionRunner` and replaces
it with the `PredictionService` added in the previous commit.

This also changes the behaviour of cog in the event of a fatal
prediction exception: if `await_explicit_shutdown` is set, then rather
than triggering instant shutdown, we will poison the healthcheck with a
new DEFUNCT health status.

Co-Authored-By: F <f@replicate.com>